refactor: extract skip scheduling engine into content/skipScheduler.ts
Extract 22 skip-scheduling functions from content.ts into a new
skipScheduler module with dependency injection via initSkipScheduler().

Exported functions: startSponsorSchedule, cancelSponsorSchedule,
skipToTime, getVirtualTime, updateVirtualTime, updateWaitingTime,
clearWaitingTime, startSkipScheduleCheckingForStartSponsors,
shouldAutoSkip, shouldSkip, unskipSponsorTime, reskipSponsorTime,
previewTime, inMuteSegment, isSegmentMarkedNearCurrentTime,
incorrectVideoCheck, createSkipNotice, createAdvanceSkipNotice.

Internal (non-exported) helpers: waitForNextTimeChange,
getNextSkipIndex, getLatestEndTimeIndex, getStartTimes,
sendTelemetryAndCount.

Dependencies injected via SkipSchedulerDeps interface:
- skipNoticeContentContainer
- updateActiveSegment

A local Utils instance is created inside the module for
getCategorySelection() calls.

Cleaned up unused imports in content.ts.
