+++ src/content.ts
@@ -5,11 +5,8 @@ import { ContentContainer } from "./ContentContainerTypes";
 import { SkipButtonControlBar } from "./js-components/skipButtonControlBar";
 import {
     contentState,
-    endTimeSkipBuffer,
     getPageLoaded,
-    manualSkipPercentCount,
     setupPageLoadingListener,
-    skipBuffer,
 } from "./content/state";
 import { danmakuForSkip, initDanmakuSkip } from "./content/danmakuSkip";
 import {
@@ -21,16 +18,29 @@ import {
     updateActiveSegment,
     updatePreviewBar,
 } from "./content/previewBarManager";
+import {
+    cancelSponsorSchedule,
+    clearWaitingTime,
+    getVirtualTime,
+    initSkipScheduler,
+    isSegmentMarkedNearCurrentTime,
+    previewTime,
+    reskipSponsorTime,
+    skipToTime,
+    startSkipScheduleCheckingForStartSponsors,
+    startSponsorSchedule,
+    unskipSponsorTime,
+    updateVirtualTime,
+    updateWaitingTime,
+} from "./content/skipScheduler";
 import { initMessageHandler, setupMessageListener } from "./content/messageHandler";
 import { addHotkeyListener, initHotkeyHandler, seekFrameByKeyPressListener } from "./content/hotkeyHandler";
 import { VoteResponse } from "./messageTypes";
-import advanceSkipNotice from "./render/advanceSkipNotice";
 import { CategoryPill } from "./render/CategoryPill";
 import { DescriptionPortPill } from "./render/DescriptionPortPill";
 import { CommentListener, DynamicListener } from "./render/DynamicAndCommentSponsorBlock";
 import { setMessageNotice, showMessage } from "./render/MessageNotice";
 import { PlayerButton } from "./render/PlayerButton";
-import SkipNotice from "./render/SkipNotice";
 import SubmissionNotice from "./render/SubmissionNotice";
 import { getPortVideoByHash, postPortVideo, postPortVideoVote, updatePortedSegments } from "./requests/portVideo";
 import { asyncRequestToServer } from "./requests/requests";
@@ -42,39 +52,31 @@ import {
     ActionType,
     BVID,
     Category,
-    CategorySkipOption,
     ChannelIDInfo,
     ChannelIDStatus,
     NewVideoID,
     PageType,
     PortVideo,
-    ScheduledTime,
     SegmentUUID,
-    SkipToTimeParams,
     SponsorHideType,
     SponsorSourceType,
     SponsorTime,
     YTID,
 } from "./types";
 import Utils from "./utils";
-import { isFirefox, isFirefoxOrSafari, isSafari, waitFor } from "./utils/";
+import { waitFor } from "./utils/";
 import { AnimationUtils } from "./utils/animationUtils";
 import { addCleanupListener, cleanPage } from "./utils/cleanup";
 import { defaultPreviewTime } from "./utils/constants";
-import { parseTargetTimeFromDanmaku } from "./utils/danmakusUtils";
 import { durationEquals } from "./utils/duraionUtils";
 import { getErrorMessage, getFormattedTime } from "./utils/formating";
 import { GenericUtils } from "./utils/genericUtils";
 import { getHash, getVideoIDHash, HashedValue } from "./utils/hash";
 import { getCidMapFromWindow } from "./utils/injectedScriptMessageUtils";
 import { logDebug } from "./utils/logger";
-import { getControls, getHashParams, getProgressBar, isPlayingPlaylist } from "./utils/pageUtils";
-import { getBilibiliVideoID } from "./utils/parseVideoID";
+import { getControls, getHashParams, getProgressBar } from "./utils/pageUtils";
 import { generateUserID } from "./utils/setup";
-import { getStartTimeFromUrl } from "./utils/urlParser";
 import {
-    checkIfNewVideoID,
-    checkVideoIDChange,
     detectPageType,
     getBvID,
     getChannelIDInfo,
@@ -169,6 +171,11 @@ const skipNoticeContentContainer: ContentContainer = () => ({
     getRealCurrentTime: getRealCurrentTime,
     lockedCategories: contentState.lockedCategories,
     channelIDInfo: getChannelIDInfo(),
+});
+
+initSkipScheduler({
+    skipNoticeContentContainer,
+    updateActiveSegment,
 });
 
 initMessageHandler({
@@ -293,345 +300,6 @@ function videoOnReadyListener(): void {
     updateVisibilityOfPlayerControlsButton();
 }
 
-function cancelSponsorSchedule(): void {
-    logDebug("Pausing skipping");
-
-    if (contentState.currentSkipSchedule !== null) {
-        clearTimeout(contentState.currentSkipSchedule);
-        contentState.currentSkipSchedule = null;
-    }
-
-    if (contentState.currentSkipInterval !== null) {
-        clearInterval(contentState.currentSkipInterval);
-        contentState.currentSkipInterval = null;
-    }
-
-    if (contentState.currentadvanceSkipSchedule !== null) {
-        clearInterval(contentState.currentadvanceSkipSchedule);
-        contentState.currentadvanceSkipSchedule = null;
-    }
-}
-
-/**
- * @param currentTime Optional if you don't want to use the actual current time
- */
-async function startSponsorSchedule(
-    includeIntersectingSegments = false,
-    currentTime?: number,
-    includeNonIntersectingSegments = true
-): Promise<void> {
-    cancelSponsorSchedule();
-
-    // Give up if video changed, and trigger a videoID change if so
-    if (await checkIfNewVideoID()) {
-        return;
-    }
-
-    const video = getVideo();
-    logDebug(`Considering to start skipping: ${!video}, ${video?.paused}`);
-    if (!video) return;
-    if (currentTime === undefined || currentTime === null) {
-        currentTime = getVirtualTime();
-    }
-    clearWaitingTime();
-
-    updateActiveSegment(currentTime);
-
-    if (video.paused || (video.currentTime >= video.duration - 0.01 && video.duration > 1)) return;
-    const skipInfo = getNextSkipIndex(currentTime, includeIntersectingSegments, includeNonIntersectingSegments);
-
-    const currentSkip = skipInfo.array[skipInfo.index];
-    const skipTime: number[] = [currentSkip?.scheduledTime, skipInfo.array[skipInfo.endIndex]?.segment[1]];
-    const timeUntilSponsor = skipTime?.[0] - currentTime;
-    const videoID = getVideoID();
-
-    if (
-        contentState.videoMuted &&
-        !inMuteSegment(
-            currentTime,
-            skipInfo.index !== -1 && timeUntilSponsor < skipBuffer && shouldAutoSkip(currentSkip)
-        )
-    ) {
-        video.muted = false;
-        contentState.videoMuted = false;
-
-        for (const notice of contentState.skipNotices) {
-            // So that the notice can hide buttons
-            notice.unmutedListener(currentTime);
-        }
-    }
-
-    logDebug(`Ready to start skipping: ${skipInfo.index} at ${currentTime}`);
-    if (skipInfo.index === -1) return;
-
-    if (
-        Config.config.disableSkipping ||
-        contentState.channelWhitelisted ||
-        (getChannelIDInfo().status === ChannelIDStatus.Fetching && Config.config.forceChannelCheck)
-    ) {
-        return;
-    }
-
-    if (await incorrectVideoCheck()) return;
-
-    // Find all indexes in between the start and end
-    let skippingSegments = [skipInfo.array[skipInfo.index]];
-    if (skipInfo.index !== skipInfo.endIndex) {
-        skippingSegments = [];
-
-        for (const segment of skipInfo.array) {
-            if (
-                shouldAutoSkip(segment) &&
-                segment.segment[0] >= skipTime[0] &&
-                segment.segment[1] <= skipTime[1] &&
-                segment.segment[0] === segment.scheduledTime
-            ) {
-                // Don't include artifical scheduled segments (end times for mutes)
-                skippingSegments.push(segment);
-            }
-        }
-    }
-
-    logDebug(
-        `Next step in starting skipping: ${!shouldSkip(currentSkip)}, ${!contentState.sponsorTimesSubmitting?.some(
-            (segment) => segment.segment === currentSkip.segment
-        )}`
-    );
-
-    const skippingFunction = async (forceVideoTime?: number) => {
-        let forcedSkipTime: number = null;
-        let forcedIncludeIntersectingSegments = false;
-        let forcedIncludeNonIntersectingSegments = true;
-
-        if (await incorrectVideoCheck(videoID, currentSkip)) return;
-        forceVideoTime ||= Math.max(getVideo().currentTime, getVirtualTime());
-
-        if (
-            shouldSkip(currentSkip) ||
-            contentState.sponsorTimesSubmitting?.some((segment) => segment.segment === currentSkip.segment)
-        ) {
-            if (forceVideoTime >= skipTime[0] - skipBuffer && forceVideoTime < skipTime[1]) {
-                skipToTime({
-                    v: getVideo(),
-                    skipTime,
-                    skippingSegments,
-                    openNotice: skipInfo.openNotice,
-                });
-
-                // These are segments that start at the exact same time but need seperate notices
-                for (const extra of skipInfo.extraIndexes) {
-                    const extraSkip = skipInfo.array[extra];
-                    if (shouldSkip(extraSkip)) {
-                        skipToTime({
-                            v: getVideo(),
-                            skipTime: [extraSkip.scheduledTime, extraSkip.segment[1]],
-                            skippingSegments: [extraSkip],
-                            openNotice: skipInfo.openNotice,
-                        });
-                    }
-                }
-
-                if (
-                    utils.getCategorySelection(currentSkip.category)?.option === CategorySkipOption.ManualSkip ||
-                    currentSkip.actionType === ActionType.Mute
-                ) {
-                    forcedSkipTime = skipTime[0] + 0.001;
-                } else {
-                    forcedSkipTime = skipTime[1];
-                    forcedIncludeNonIntersectingSegments = false;
-
-                    // Only if not at the end of the video
-                    if (Math.abs(skipTime[1] - getVideo().duration) > endTimeSkipBuffer) {
-                        forcedIncludeIntersectingSegments = true;
-                    }
-                }
-            } else {
-                forcedSkipTime = forceVideoTime + 0.001;
-            }
-        } else {
-            forcedSkipTime = forceVideoTime + 0.001;
-        }
-
-        // Don't pretend to be earlier than we are, could result in loops
-        if (forcedSkipTime !== null && forceVideoTime > forcedSkipTime) {
-            forcedSkipTime = forceVideoTime;
-        }
-
-        startSponsorSchedule(forcedIncludeIntersectingSegments, forcedSkipTime, forcedIncludeNonIntersectingSegments);
-    };
-
-    if (timeUntilSponsor < skipBuffer) {
-        await skippingFunction(currentTime);
-    } else {
-        let delayTime = (timeUntilSponsor * 1000) / getVideo().playbackRate;
-        if (delayTime < (isFirefox() ? 750 : 300) && shouldAutoSkip(skippingSegments[0])) {
-            let forceStartIntervalTime: number | null = null;
-            if (isFirefox() && delayTime > 300) {
-                forceStartIntervalTime = await waitForNextTimeChange();
-            }
-
-            // Use interval instead of timeout near the end to combat imprecise video time
-            const startIntervalTime = forceStartIntervalTime || performance.now();
-            const startVideoTime = Math.max(currentTime, getVideo().currentTime);
-            delayTime = (skipTime?.[0] - startVideoTime) * 1000 * (1 / getVideo().playbackRate);
-
-            let startWaitingForReportedTimeToChange = true;
-            const reportedVideoTimeAtStart = getVideo().currentTime;
-            logDebug(`Starting setInterval skipping ${getVideo().currentTime} to skip at ${skipTime[0]}`);
-
-            if (contentState.currentSkipInterval !== null) clearInterval(contentState.currentSkipInterval);
-            contentState.currentSkipInterval = setInterval(() => {
-                // Estimate delay, but only take the current time right after a change
-                // Current time remains the same for many "frames" on Firefox
-                if (
-                    isFirefoxOrSafari() &&
-                    !contentState.lastKnownVideoTime.fromPause &&
-                    startWaitingForReportedTimeToChange &&
-                    reportedVideoTimeAtStart !== getVideo().currentTime
-                ) {
-                    startWaitingForReportedTimeToChange = false;
-                    const delay = getVirtualTime() - getVideo().currentTime;
-                    if (delay > 0) contentState.lastKnownVideoTime.approximateDelay = delay;
-                }
-
-                const intervalDuration = performance.now() - startIntervalTime;
-                if (intervalDuration + skipBuffer * 1000 >= delayTime || getVideo().currentTime >= skipTime[0]) {
-                    clearInterval(contentState.currentSkipInterval);
-                    if (!isFirefoxOrSafari() && !getVideo().muted && !inMuteSegment(getVideo().currentTime, true)) {
-                        // Workaround for more accurate skipping on Chromium
-                        getVideo().muted = true;
-                        getVideo().muted = false;
-                    }
-
-                    skippingFunction(
-                        Math.max(
-                            getVideo().currentTime,
-                            startVideoTime + (getVideo().playbackRate * Math.max(delayTime, intervalDuration)) / 1000
-                        )
-                    );
-                }
-            }, 0);
-        } else {
-            logDebug(`Starting timeout to skip ${getVideo().currentTime} to skip at ${skipTime[0]}`);
-
-            const offset = isFirefoxOrSafari() && !isSafari() ? 600 : 150;
-            // Schedule for right before to be more precise than normal timeout
-            const offsetDelayTime = Math.max(0, delayTime - offset);
-            contentState.currentSkipSchedule = setTimeout(skippingFunction, offsetDelayTime);
-
-            if (
-                Config.config.advanceSkipNotice &&
-                Config.config.skipNoticeDurationBefore > 0 &&
-                getVideo().currentTime < skippingSegments[0].segment[0] &&
-                !contentState.sponsorTimesSubmitting?.some((segment) => segment.segment === currentSkip.segment) &&
-                [ActionType.Skip, ActionType.Mute].includes(skippingSegments[0].actionType) &&
-                shouldAutoSkip(skippingSegments[0]) &&
-                !getVideo()?.paused
-            ) {
-                const maxPopupTime = Config.config.skipNoticeDurationBefore * 1000;
-                const timeUntilPopup = Math.max(0, offsetDelayTime - maxPopupTime);
-                const autoSkip = shouldAutoSkip(skippingSegments[0]);
-
-                if (contentState.currentadvanceSkipSchedule) clearTimeout(contentState.currentadvanceSkipSchedule);
-                contentState.currentadvanceSkipSchedule = setTimeout(() => {
-                    createAdvanceSkipNotice([skippingSegments[0]], skipTime[0], autoSkip, false);
-                    sessionStorage.setItem("SKIPPING", "true");
-                }, timeUntilPopup);
-            }
-        }
-    }
-}
-
-
-/**
- * Used on Firefox only, waits for the next animation frame until
- * the video time has changed
- */
-function waitForNextTimeChange(): Promise<DOMHighResTimeStamp | null> {
-    return new Promise((resolve) => {
-        getVideo().addEventListener("timeupdate", () => resolve(performance.now()), { once: true });
-    });
-}
-
-function getVirtualTime(): number {
-    const virtualTime =
-        contentState.lastTimeFromWaitingEvent ??
-        (contentState.lastKnownVideoTime.videoTime !== null
-            ? ((performance.now() - contentState.lastKnownVideoTime.preciseTime) * getVideo().playbackRate) / 1000 +
-            contentState.lastKnownVideoTime.videoTime
-            : null);
-
-    if (
-        Config.config.useVirtualTime &&
-        !isSafari() &&
-        virtualTime &&
-        Math.abs(virtualTime - getVideo().currentTime) < 0.2 &&
-        getVideo().currentTime !== 0
-    ) {
-        return Math.max(virtualTime, getVideo().currentTime);
-    } else {
-        return getVideo().currentTime;
-    }
-}
-
-function inMuteSegment(currentTime: number, includeOverlap: boolean): boolean {
-    const checkFunction = (segment) =>
-        segment.actionType === ActionType.Mute &&
-        segment.hidden === SponsorHideType.Visible &&
-        segment.segment[0] <= currentTime &&
-        (segment.segment[1] > currentTime || (includeOverlap && segment.segment[1] + 0.02 > currentTime));
-    return contentState.sponsorTimes?.some(checkFunction) || contentState.sponsorTimesSubmitting.some(checkFunction);
-}
-
-function isSegmentMarkedNearCurrentTime(currentTime: number, range: number = 5): boolean {
-    const lowerBound = currentTime - range;
-    const upperBound = currentTime + range;
-
-    return contentState.sponsorTimes?.some((sponsorTime) => {
-        const {
-            segment: [startTime, endTime],
-        } = sponsorTime;
-        return startTime <= upperBound && endTime >= lowerBound;
-    });
-}
-
-/**
- * This makes sure the videoID is still correct and if the sponsorTime is included
- */
-async function incorrectVideoCheck(videoID?: string, sponsorTime?: SponsorTime): Promise<boolean> {
-    const currentVideoID = await getBilibiliVideoID();
-    const recordedVideoID = videoID || getVideoID();
-    if (
-        currentVideoID !== recordedVideoID ||
-        (sponsorTime &&
-            (!contentState.sponsorTimes ||
-                !contentState.sponsorTimes?.some(
-                    (time) => time.segment[0] === sponsorTime.segment[0] && time.segment[1] === sponsorTime.segment[1]
-                )) &&
-            !contentState.sponsorTimesSubmitting.some(
-                (time) => time.segment[0] === sponsorTime.segment[0] && time.segment[1] === sponsorTime.segment[1]
-            ))
-    ) {
-        // Something has really gone wrong
-        console.error("[SponsorBlock] The videoID recorded when trying to skip is different than what it should be.");
-        console.error("[SponsorBlock] VideoID recorded: " + recordedVideoID + ". Actual VideoID: " + currentVideoID);
-        console.error(
-            "[SponsorBlock] SponsorTime",
-            sponsorTime,
-            "sponsorTimes",
-            contentState.sponsorTimes,
-            "sponsorTimesSubmitting",
-            contentState.sponsorTimesSubmitting
-        );
-
-        // Video ID change occured
-        checkVideoIDChange();
-
-        return true;
-    } else {
-        return false;
-    }
-}
 
 let playbackRateCheckInterval: NodeJS.Timeout | null = null;
 let lastPlaybackSpeed = 1;
@@ -829,56 +497,6 @@ function setupVideoListeners(video: HTMLVideoElement) {
     setupVideoListenersFirstTime = false;
 }
 
-function updateVirtualTime() {
-    if (contentState.currentVirtualTimeInterval) clearInterval(contentState.currentVirtualTimeInterval);
-
-    contentState.lastKnownVideoTime.videoTime = getVideo().currentTime;
-    contentState.lastKnownVideoTime.preciseTime = performance.now();
-
-    // If on Firefox, wait for the second time change (time remains fixed for many "frames" for privacy reasons)
-    if (isFirefoxOrSafari()) {
-        let count = 0;
-        let rawCount = 0;
-        let lastTime = contentState.lastKnownVideoTime.videoTime;
-        let lastPerformanceTime = performance.now();
-
-        contentState.currentVirtualTimeInterval = setInterval(() => {
-            const frameTime = performance.now() - lastPerformanceTime;
-            if (lastTime !== getVideo().currentTime) {
-                rawCount++;
-
-                // If there is lag, give it another shot at finding a good change time
-                if (frameTime < 20 || rawCount > 30) {
-                    count++;
-                }
-                lastTime = getVideo().currentTime;
-            }
-
-            if (count > 1) {
-                const delay =
-                    contentState.lastKnownVideoTime.fromPause && contentState.lastKnownVideoTime.approximateDelay
-                        ? contentState.lastKnownVideoTime.approximateDelay
-                        : 0;
-
-                contentState.lastKnownVideoTime.videoTime = getVideo().currentTime + delay;
-                contentState.lastKnownVideoTime.preciseTime = performance.now();
-
-                clearInterval(contentState.currentVirtualTimeInterval);
-                contentState.currentVirtualTimeInterval = null;
-            }
-
-            lastPerformanceTime = performance.now();
-        }, 1);
-    }
-}
-
-function updateWaitingTime(): void {
-    contentState.lastTimeFromWaitingEvent = getVideo().currentTime;
-}
-
-function clearWaitingTime(): void {
-    contentState.lastTimeFromWaitingEvent = null;
-}
 
 function setupSkipButtonControlBar() {
     if (!contentState.skipButtonControlBar) {
@@ -1110,82 +728,6 @@ async function lockedCategoriesLookup(): Promise<void> {
 }
 
 
-/**
- * Only should be used when it is okay to skip a sponsor when in the middle of it
- *
- * Ex. When segments are first loaded
- */
-function startSkipScheduleCheckingForStartSponsors() {
-    // switchingVideos is ignored in Safari due to event fire order. See #1142
-    if ((!contentState.switchingVideos || isSafari()) && contentState.sponsorTimes) {
-        // See if there are any starting sponsors
-        let startingSegmentTime = getStartTimeFromUrl(document.URL) || -1;
-        let found = false;
-        for (const time of contentState.sponsorTimes) {
-            if (
-                time.segment[0] <= getVideo().currentTime &&
-                time.segment[0] > startingSegmentTime &&
-                time.segment[1] > getVideo().currentTime &&
-                time.actionType !== ActionType.Poi
-            ) {
-                startingSegmentTime = time.segment[0];
-                found = true;
-                break;
-            }
-        }
-        if (!found) {
-            for (const time of contentState.sponsorTimesSubmitting) {
-                if (
-                    time.segment[0] <= getVideo().currentTime &&
-                    time.segment[0] > startingSegmentTime &&
-                    time.segment[1] > getVideo().currentTime &&
-                    time.actionType !== ActionType.Poi
-                ) {
-                    startingSegmentTime = time.segment[0];
-                    found = true;
-                    break;
-                }
-            }
-        }
-
-        // For highlight category
-        const poiSegments = contentState.sponsorTimes
-            .filter(
-                (time) =>
-                    time.segment[1] > getVideo().currentTime &&
-                    time.actionType === ActionType.Poi &&
-                    time.hidden === SponsorHideType.Visible
-            )
-            .sort((a, b) => b.segment[0] - a.segment[0]);
-        for (const time of poiSegments) {
-            const skipOption = utils.getCategorySelection(time.category)?.option;
-            if (skipOption !== CategorySkipOption.ShowOverlay) {
-                skipToTime({
-                    v: getVideo(),
-                    skipTime: time.segment,
-                    skippingSegments: [time],
-                    openNotice: true,
-                    unskipTime: getVideo().currentTime,
-                });
-                if (skipOption === CategorySkipOption.AutoSkip) break;
-            }
-        }
-
-        const fullVideoSegment = contentState.sponsorTimes.filter((time) => time.actionType === ActionType.Full)[0];
-        if (fullVideoSegment) {
-            waitFor(() => contentState.categoryPill).then(() => {
-                contentState.categoryPill?.setSegment(fullVideoSegment);
-            });
-        }
-
-        if (startingSegmentTime !== -1) {
-            startSponsorSchedule(undefined, startingSegmentTime);
-        } else {
-            startSponsorSchedule();
-        }
-    }
-}
-
 //checks if this channel is whitelisted, should be done only after the channelID has been loaded
 async function channelIDChange(channelIDInfo: ChannelIDInfo) {
     const whitelistedChannels = Config.config.whitelistedChannels;
@@ -1222,495 +764,6 @@ function videoElementChange(newVideo: boolean, video: HTMLVideoElement): void {
     });
 }
 
-
-/**
- * Returns info about the next upcoming sponsor skip
- */
-function getNextSkipIndex(
-    currentTime: number,
-    includeIntersectingSegments: boolean,
-    includeNonIntersectingSegments: boolean
-): { array: ScheduledTime[]; index: number; endIndex: number; extraIndexes: number[]; openNotice: boolean } {
-    const autoSkipSorter = (segment: ScheduledTime) => {
-        const skipOption = utils.getCategorySelection(segment.category)?.option;
-        if (
-            (skipOption === CategorySkipOption.AutoSkip || shouldAutoSkip(segment)) &&
-            segment.actionType === ActionType.Skip
-        ) {
-            return 0;
-        } else if (skipOption !== CategorySkipOption.ShowOverlay) {
-            return 1;
-        } else {
-            return 2;
-        }
-    };
-
-    const { includedTimes: submittedArray, scheduledTimes: sponsorStartTimes } = getStartTimes(
-        contentState.sponsorTimes,
-        includeIntersectingSegments,
-        includeNonIntersectingSegments
-    );
-    const { scheduledTimes: sponsorStartTimesAfterCurrentTime } = getStartTimes(
-        contentState.sponsorTimes,
-        includeIntersectingSegments,
-        includeNonIntersectingSegments,
-        currentTime,
-        true
-    );
-
-    // This is an array in-case multiple segments have the exact same start time
-    const minSponsorTimeIndexes = GenericUtils.indexesOf(
-        sponsorStartTimes,
-        Math.min(...sponsorStartTimesAfterCurrentTime)
-    );
-    // Find auto skipping segments if possible, sort by duration otherwise
-    const minSponsorTimeIndex =
-        minSponsorTimeIndexes.sort(
-            (a, b) =>
-                autoSkipSorter(submittedArray[a]) - autoSkipSorter(submittedArray[b]) ||
-                submittedArray[a].segment[1] -
-                submittedArray[a].segment[0] -
-                (submittedArray[b].segment[1] - submittedArray[b].segment[0])
-        )[0] ?? -1;
-    // Store extra indexes for the non-auto skipping segments if others occur at the exact same start time
-    const extraIndexes = minSponsorTimeIndexes.filter(
-        (i) => i !== minSponsorTimeIndex && autoSkipSorter(submittedArray[i]) !== 0
-    );
-
-    const endTimeIndex = getLatestEndTimeIndex(submittedArray, minSponsorTimeIndex);
-
-    const { includedTimes: unsubmittedArray, scheduledTimes: unsubmittedSponsorStartTimes } = getStartTimes(
-        contentState.sponsorTimesSubmitting,
-        includeIntersectingSegments,
-        includeNonIntersectingSegments
-    );
-    const { scheduledTimes: unsubmittedSponsorStartTimesAfterCurrentTime } = getStartTimes(
-        contentState.sponsorTimesSubmitting,
-        includeIntersectingSegments,
-        includeNonIntersectingSegments,
-        currentTime,
-        false
-    );
-
-    const minUnsubmittedSponsorTimeIndex = unsubmittedSponsorStartTimes.indexOf(
-        Math.min(...unsubmittedSponsorStartTimesAfterCurrentTime)
-    );
-    const previewEndTimeIndex = getLatestEndTimeIndex(unsubmittedArray, minUnsubmittedSponsorTimeIndex);
-
-    if (
-        (minUnsubmittedSponsorTimeIndex === -1 && minSponsorTimeIndex !== -1) ||
-        sponsorStartTimes[minSponsorTimeIndex] < unsubmittedSponsorStartTimes[minUnsubmittedSponsorTimeIndex]
-    ) {
-        return {
-            array: submittedArray,
-            index: minSponsorTimeIndex,
-            endIndex: endTimeIndex,
-            extraIndexes, // Segments at same time that need seperate notices
-            openNotice: true,
-        };
-    } else {
-        return {
-            array: unsubmittedArray,
-            index: minUnsubmittedSponsorTimeIndex,
-            endIndex: previewEndTimeIndex,
-            extraIndexes: [], // No manual things for unsubmitted
-            openNotice: false,
-        };
-    }
-}
-
-/**
- * This returns index if the skip option is not AutoSkip
- *
- * Finds the last endTime that occurs in a segment that the given
- * segment skips into that is part of an AutoSkip category.
- *
- * Used to find where a segment should truely skip to if there are intersecting submissions due to
- * them having different categories.
- *
- * @param sponsorTimes
- * @param index Index of the given sponsor
- * @param hideHiddenSponsors
- */
-function getLatestEndTimeIndex(sponsorTimes: SponsorTime[], index: number, hideHiddenSponsors = true): number {
-    // Only combine segments for AutoSkip
-    if (index == -1 || !shouldAutoSkip(sponsorTimes[index]) || sponsorTimes[index].actionType !== ActionType.Skip) {
-        return index;
-    }
-
-    // Default to the normal endTime
-    let latestEndTimeIndex = index;
-
-    for (let i = 0; i < sponsorTimes?.length; i++) {
-        const currentSegment = sponsorTimes[i].segment;
-        const latestEndTime = sponsorTimes[latestEndTimeIndex].segment[1];
-
-        if (
-            currentSegment[0] - skipBuffer <= latestEndTime &&
-            currentSegment[1] > latestEndTime &&
-            (!hideHiddenSponsors || sponsorTimes[i].hidden === SponsorHideType.Visible) &&
-            shouldAutoSkip(sponsorTimes[i]) &&
-            sponsorTimes[i].actionType === ActionType.Skip
-        ) {
-            // Overlapping segment
-            latestEndTimeIndex = i;
-        }
-    }
-
-    // Keep going if required
-    if (latestEndTimeIndex !== index) {
-        latestEndTimeIndex = getLatestEndTimeIndex(sponsorTimes, latestEndTimeIndex, hideHiddenSponsors);
-    }
-
-    return latestEndTimeIndex;
-}
-
-/**
- * Gets just the start times from a sponsor times array.
- * Optionally specify a minimum
- *
- * @param sponsorTimes
- * @param minimum
- * @param hideHiddenSponsors
- * @param includeIntersectingSegments If true, it will include segments that start before
- *  the current time, but end after
- */
-function getStartTimes(
-    sponsorTimes: SponsorTime[],
-    includeIntersectingSegments: boolean,
-    includeNonIntersectingSegments: boolean,
-    minimum?: number,
-    hideHiddenSponsors = false
-): { includedTimes: ScheduledTime[]; scheduledTimes: number[] } {
-    if (!sponsorTimes) return { includedTimes: [], scheduledTimes: [] };
-
-    const includedTimes: ScheduledTime[] = [];
-    const scheduledTimes: number[] = [];
-
-    const shouldIncludeTime = (segment: ScheduledTime) =>
-        (minimum === undefined ||
-            (includeNonIntersectingSegments && segment.scheduledTime >= minimum) ||
-            (includeIntersectingSegments &&
-                segment.scheduledTime < minimum &&
-                segment.segment[1] > minimum &&
-                shouldSkip(segment))) && // Only include intersecting skippable segments
-        (!hideHiddenSponsors || segment.hidden === SponsorHideType.Visible) &&
-        segment.segment.length === 2 &&
-        segment.actionType !== ActionType.Poi &&
-        segment.actionType !== ActionType.Full;
-
-    const possibleTimes = sponsorTimes.map((sponsorTime) => ({
-        ...sponsorTime,
-        scheduledTime: sponsorTime.segment[0],
-    }));
-
-    // Schedule at the end time to know when to unmute and remove title from seek bar
-    sponsorTimes.forEach((sponsorTime) => {
-        if (
-            !possibleTimes.some((time) => sponsorTime.segment[1] === time.scheduledTime && shouldIncludeTime(time)) &&
-            (minimum === undefined || sponsorTime.segment[1] > minimum)
-        ) {
-            possibleTimes.push({
-                ...sponsorTime,
-                scheduledTime: sponsorTime.segment[1],
-            });
-        }
-    });
-
-    for (let i = 0; i < possibleTimes.length; i++) {
-        if (shouldIncludeTime(possibleTimes[i])) {
-            scheduledTimes.push(possibleTimes[i].scheduledTime);
-            includedTimes.push(possibleTimes[i]);
-        }
-    }
-
-    return { includedTimes, scheduledTimes };
-}
-
-/**
- * Skip to exact time in a video and autoskips
- *
- * @param time
- */
-function previewTime(time: number, unpause = true) {
-    contentState.previewedSegment = true;
-    getVideo().currentTime = time;
-
-    // Unpause the video if needed
-    if (unpause && getVideo().paused) {
-        getVideo().play();
-    }
-}
-
-//send telemetry and count skip
-function sendTelemetryAndCount(skippingSegments: SponsorTime[], secondsSkipped: number, fullSkip: boolean) {
-    for (const segment of skippingSegments) {
-        if (!contentState.previewedSegment && contentState.sponsorTimesSubmitting.some((s) => s.segment === segment.segment)) {
-            // Count that as a previewed segment
-            contentState.previewedSegment = true;
-        }
-    }
-
-    if (
-        !Config.config.trackViewCount ||
-        (!Config.config.trackViewCountInPrivate && chrome.extension.inIncognitoContext)
-    )
-        return;
-
-    let counted = false;
-    for (const segment of skippingSegments) {
-        const index = contentState.sponsorTimes?.findIndex((s) => s.segment === segment.segment);
-        if (index !== -1 && !contentState.sponsorSkipped[index]) {
-            contentState.sponsorSkipped[index] = true;
-            if (!counted) {
-                Config.config.minutesSaved = Config.config.minutesSaved + secondsSkipped / 60;
-                Config.config.skipCount = Config.config.skipCount + 1;
-                counted = true;
-            }
-
-            if (fullSkip) asyncRequestToServer("POST", "/api/viewedVideoSponsorTime?UUID=" + segment.UUID);
-        }
-    }
-}
-
-//skip from the start time to the end time for a certain index sponsor time
-function skipToTime({ v, skipTime, skippingSegments, openNotice, forceAutoSkip, unskipTime }: SkipToTimeParams): void {
-    if (Config.config.disableSkipping) return;
-
-    let autoSkip: boolean;
-    if (sessionStorage.getItem("SKIPPING") === "false") {
-        sessionStorage.setItem("SKIPPING", "null");
-        autoSkip = false;
-    } else {
-        // There will only be one submission if it is manual skip
-        autoSkip = forceAutoSkip || shouldAutoSkip(skippingSegments[0]);
-    }
-
-    const isSubmittingSegment = contentState.sponsorTimesSubmitting.some((time) => time.segment === skippingSegments[0].segment);
-
-    if ((autoSkip || isSubmittingSegment) && v.currentTime !== skipTime[1]) {
-        switch (skippingSegments[0].actionType) {
-            case ActionType.Poi:
-            case ActionType.Skip: {
-                // Fix for looped videos not working when skipping to the end #426
-                // for some reason you also can't skip to 1 second before the end
-                if (v.loop && v.duration > 1 && skipTime[1] >= v.duration - 1) {
-                    v.currentTime = 0;
-                } else if (
-                    v.duration > 1 &&
-                    skipTime[1] >= v.duration &&
-                    (navigator.vendor === "Apple Computer, Inc." || isPlayingPlaylist())
-                ) {
-                    // MacOS will loop otherwise #1027
-                    // Sometimes playlists loop too #1804
-                    v.currentTime = v.duration - 0.001;
-                } else if (
-                    v.duration > 1 &&
-                    Math.abs(skipTime[1] - v.duration) < endTimeSkipBuffer &&
-                    isFirefoxOrSafari() &&
-                    !isSafari()
-                ) {
-                    v.currentTime = v.duration;
-                } else {
-                    if (inMuteSegment(skipTime[1], true)) {
-                        // Make sure not to mute if skipping into a mute segment
-                        v.muted = true;
-                        contentState.videoMuted = true;
-                    }
-
-                    v.currentTime = skipTime[1];
-                }
-
-                break;
-            }
-            case ActionType.Mute: {
-                if (!v.muted) {
-                    v.muted = true;
-                    contentState.videoMuted = true;
-                }
-                break;
-            }
-        }
-    }
-
-    if (autoSkip && Config.config.audioNotificationOnSkip && !isSubmittingSegment && !getVideo()?.muted) {
-        const beep = new Audio(chrome.runtime.getURL("icons/beep.ogg"));
-        beep.volume = getVideo().volume * 0.1;
-        const oldMetadata = navigator.mediaSession.metadata;
-        beep.play();
-        beep.addEventListener("ended", () => {
-            navigator.mediaSession.metadata = null;
-            setTimeout(() => {
-                navigator.mediaSession.metadata = oldMetadata;
-                beep.remove();
-            });
-        });
-    }
-
-    if (!autoSkip && skippingSegments.length === 1 && skippingSegments[0].actionType === ActionType.Poi) {
-        waitFor(() => contentState.skipButtonControlBar).then(() => {
-            contentState.skipButtonControlBar.enable(skippingSegments[0]);
-            if (Config.config.skipKeybind == null) contentState.skipButtonControlBar.setShowKeybindHint(false);
-
-            contentState.activeSkipKeybindElement?.setShowKeybindHint(false);
-            contentState.activeSkipKeybindElement = contentState.skipButtonControlBar;
-        });
-    } else {
-        if (openNotice) {
-            //send out the message saying that a sponsor message was skipped
-            if (!Config.config.dontShowNotice || !autoSkip) {
-                createSkipNotice(skippingSegments, autoSkip, unskipTime, false);
-            } else if (autoSkip) {
-                contentState.activeSkipKeybindElement?.setShowKeybindHint(false);
-                contentState.activeSkipKeybindElement = {
-                    setShowKeybindHint: () => { },
-                    toggleSkip: () => {
-                        createSkipNotice(skippingSegments, autoSkip, unskipTime, true);
-
-                        unskipSponsorTime(skippingSegments[0], unskipTime);
-                    },
-                };
-            }
-        }
-    }
-
-    //send telemetry that a this sponsor was skipped
-    if (autoSkip || isSubmittingSegment) sendTelemetryAndCount(skippingSegments, skipTime[1] - skipTime[0], true);
-}
-
-function createSkipNotice(
-    skippingSegments: SponsorTime[],
-    autoSkip: boolean,
-    unskipTime: number,
-    startReskip: boolean
-) {
-    for (const skipNotice of contentState.skipNotices) {
-        if (
-            skippingSegments.length === skipNotice.segments.length &&
-            skippingSegments.every((segment) => skipNotice.segments.some((s) => s.UUID === segment.UUID))
-        ) {
-            // Skip notice already exists
-            return;
-        }
-    }
-
-    const advanceSkipNoticeShow = !!contentState.advanceSkipNotices;
-    const newSkipNotice = new SkipNotice(
-        skippingSegments,
-        autoSkip,
-        skipNoticeContentContainer,
-        () => {
-            contentState.advanceSkipNotices?.close();
-            contentState.advanceSkipNotices = null;
-        },
-        unskipTime,
-        startReskip,
-        advanceSkipNoticeShow
-    );
-    if (Config.config.skipKeybind == null) newSkipNotice.setShowKeybindHint(false);
-    contentState.skipNotices.push(newSkipNotice);
-
-    contentState.activeSkipKeybindElement?.setShowKeybindHint(false);
-    contentState.activeSkipKeybindElement = newSkipNotice;
-}
-
-function createAdvanceSkipNotice(
-    skippingSegments: SponsorTime[],
-    unskipTime: number,
-    autoSkip: boolean,
-    startReskip: boolean
-) {
-    if (contentState.advanceSkipNotices && !contentState.advanceSkipNotices.closed && contentState.advanceSkipNotices.sameNotice(skippingSegments)) {
-        return;
-    }
-
-    contentState.advanceSkipNotices?.close();
-    contentState.advanceSkipNotices = new advanceSkipNotice(
-        skippingSegments,
-        skipNoticeContentContainer,
-        unskipTime,
-        autoSkip,
-        startReskip
-    );
-    if (Config.config.skipKeybind == null) contentState.advanceSkipNotices.setShowKeybindHint(false);
-
-    contentState.activeSkipKeybindElement?.setShowKeybindHint(false);
-    contentState.activeSkipKeybindElement = contentState.advanceSkipNotices;
-}
-
-function unskipSponsorTime(segment: SponsorTime, unskipTime: number = null, forceSeek = false) {
-    if (segment.actionType === ActionType.Mute) {
-        getVideo().muted = false;
-        contentState.videoMuted = false;
-    }
-
-    if (forceSeek || segment.actionType === ActionType.Skip) {
-        //add a tiny bit of time to make sure it is not skipped again
-        getVideo().currentTime = unskipTime ?? segment.segment[0] + 0.001;
-    }
-}
-
-function reskipSponsorTime(segment: SponsorTime, forceSeek = false) {
-    if (segment.actionType === ActionType.Mute && !forceSeek) {
-        getVideo().muted = true;
-        contentState.videoMuted = true;
-    } else {
-        const skippedTime = Math.max(segment.segment[1] - getVideo().currentTime, 0);
-        const segmentDuration = segment.segment[1] - segment.segment[0];
-        const fullSkip = skippedTime / segmentDuration > manualSkipPercentCount;
-
-        getVideo().currentTime = segment.segment[1];
-        sendTelemetryAndCount([segment], skippedTime, fullSkip);
-        startSponsorSchedule(true, segment.segment[1], false);
-    }
-}
-
-function shouldAutoSkip(segment: SponsorTime): boolean {
-    // danmaku segments are controlled by config
-    if (segment.source === SponsorSourceType.Danmaku) {
-        return Config.config.enableAutoSkipDanmakuSkip;
-    }
-    // Check if manual skip on full video is disabled or there is no full video segment for this category
-    if (
-        Config.config.manualSkipOnFullVideo &&
-        contentState.sponsorTimes?.some((s) => s.category === segment.category && s.actionType === ActionType.Full)
-    ) {
-        return false;
-    }
-
-    const categoryOption = utils.getCategorySelection(segment.category)?.option;
-
-    // Check if the category is set to AutoSkip
-    if (categoryOption === CategorySkipOption.AutoSkip) {
-        return true;
-    }
-    // Check if auto skip on music videos is enabled and there's a "music_offtopic" segment, and the current segment is of type Skip
-    else if (
-        Config.config.autoSkipOnMusicVideos &&
-        contentState.sponsorTimes?.some((s) => s.category === "music_offtopic") &&
-        segment.actionType === ActionType.Skip
-    ) {
-        return true;
-    }
-    // Check if this segment is in the submitting segments
-    else if (contentState.sponsorTimesSubmitting.some((s) => s.segment === segment.segment)) {
-        return true;
-    }
-
-    // Do not auto-skip in other cases
-    return false;
-}
-
-function shouldSkip(segment: SponsorTime): boolean {
-    return (
-        (segment.actionType !== ActionType.Full &&
-            segment.source !== SponsorSourceType.YouTube &&
-            utils.getCategorySelection(segment.category)?.option !== CategorySkipOption.ShowOverlay) ||
-        (Config.config.autoSkipOnMusicVideos &&
-            contentState.sponsorTimes?.some((s) => s.category === "music_offtopic") &&
-            segment.actionType === ActionType.Skip)
-    );
-}
 
 /** Creates any missing buttons on the player and updates their visiblity. */
 async function updateVisibilityOfPlayerControlsButton(): Promise<void> {
